Moved CSS files into a single folder
